fix
correções de erros, projeto funcionando perfeitamente.
