+++ package.json
@@ -14,6 +14,7 @@
     "nodemon": "^2.0.15",
     "ts-node": "^10.9.1",
     "tslib": "^2.4.1",
+    "tsx": "^3.12.2",
     "typescript": "^4.9.4",
     "zod": "^3.20.2"
   },
