+++ src/index.ts
@@ -1,13 +1,25 @@
-import users from "./mocks/CustomerMock";
 import fastify from "fastify";
+import { z } from "zod";
+import createPerson from "./domain/user/services/CustomerValidator";
+import users from "./mocks/CustomerMock";
 
-const app = fastify();
+const routes = fastify();
 
-app.get("/users", (request, reply) => users);
+routes.get("/users", () => users);
 
-app.post("/users", (request, reply) => users);
+routes.post("/users", (request, reply) => {
+  try {
+    const user = createPerson.parse(request.body);
+    users.push(user);
+    return user;
+  } catch (error) {
+    const errorJson = JSON.stringify(error);
+    if (error instanceof z.ZodError) return reply.status(400).send(errorJson);
+    else return reply.status(500).send(errorJson);
+  }
+});
 
-app.listen({ port: 3000 }, (err, address) => {
+routes.listen({ port: 3000 }, (err, address) => {
   if (err) {
     console.error(err);
     process.exit(1);
